packs/comfyui-q-find-mask-size/x24b8f78/ComfyUI-Q_find-mask-size-24b8f7842b743056b7db29c601e80e154809cdc7/pyproject.toml
[project]
name = "comfyui-q-find-mask-size"
description = "ComfyUI custom node for finding and analyzing mask sizes in images"
version = "1.0.0"
license = { file = "LICENSE.txt" }
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0"
]

[project.urls]
Repository = "https://github.com/angree/ComfyUI-Q_find-mask-size"

[tool.comfy]
PublisherId = "angree"
DisplayName = "Q Find Mask Size"

